Remove automatic installation of dependencies via Puppet
This feature has always been a bit brittle because some dependency
definitions didn't work with Puppet and resultet in errors. There is
also a problem with incompatibilities between different Puppet and Ruby
versions. Puppet 3.x does not work with Ruby 2.2 and Puppet 4.x only works
with Ruby 2.x.

Dependencies can be installed manually in the build method of the
recipe or in one of the lifecycle hook methods like before_build by
executing the needed apt-get/yum commands via the sh() method. This is
more work for the user but is also more flexible.