Add "epoch" field to the recipe class
Adjust the Package::Version class to handle an epoch that is set in the
version string and as recipe field.

An epoch in the version string has precedence over the recipe field for
backward compatibility.

If the epoch is set in both, the version string and the epoch field, we
throw an exception because we cannot select on value automatically.

Refs #210