Introduce two new package file creation lifecycle hooks
- Add "before_package_file_create"
- Add "after_package_file_create"
- Deprecate "before_package_create"
- Deprecate "after_package_create"

The new hooks take the package filename as first parameter.

With the introduction of the package_name_format recipe field, the
package filename cannot be computed with the FPM::Package object alone
anymore. We now pass in the filename directly so consumers don't have to
compute the filename anymore.

We could have passed in the recipe object or the package name format so
consumers could compute the filename based on that, but that would
require deeper knowledge of the FPM::Package object.